requirements.txt
Django>=5.0,<6.0
uvicorn[standard]>=0.30.0,<0.31.0
psycopg[binary]>=3.1,<4.0
python-telegram-bot>=21.0,<22.0
loguru>=0.7.0
textual>=0.80.0
httpx>=0.27.0,<0.28.0
pytoniq>=0.1.40,<0.2.0
websockets>=13.0,<14.0
certifi>=2024.0.0
